Normalize test/suite name: employment skip household employment connect
